Misc.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@43798 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/DatabaseSnapshot.h
@@ -169,7 +169,7 @@ class DatabaseSnapshot
 			memcpy(ptr, &adjusted_length, sizeof(adjusted_length));
 			ptr += sizeof(USHORT);
 			memcpy(ptr, value, length);
-			offset += delta;
+			offset += (ULONG) delta;
 		}
 
 		UCHAR buffer[MAX_FORMAT_SIZE];
